# Add support for whitespace control character.