--- a/ext/liquid_c/tokenizer.c
+++ b/ext/liquid_c/tokenizer.c
@@ -48,6 +48,7 @@ static VALUE tokenizer_initialize_method(VALUE self, VALUE source, VALUE line_nu
     tokenizer->source = source;
     tokenizer->cursor = RSTRING_PTR(source);
     tokenizer->length = RSTRING_LEN(source);
+    tokenizer->trim_whitespace = 0;
     // tokenizer->line_number keeps track of the current line number or it is 0
     // to indicate that line numbers aren't being calculated
     tokenizer->line_number = RTEST(line_numbers) ? 1 : 0;
@@ -66,6 +67,7 @@ void tokenizer_next(tokenizer_t *tokenizer, token_t *token)
 
     token->str = cursor;
     token->type = TOKEN_RAW;
+    token->trim_whitespace = 0;
 
     while (cursor < last) {
         if (*cursor++ != '{')
@@ -127,6 +129,13 @@ found:
             cursor++;
         }
     }
+
+    if(token->type == TOKEN_RAW && tokenizer->trim_whitespace ) {
+        token->trim_whitespace = 1;
+    }
+
+    tokenizer->trim_whitespace = ((token->type == TOKEN_VARIABLE || token->type == TOKEN_TAG) && token->str[token->length - 3] == '-');
+
 }
 
 static VALUE tokenizer_shift_method(VALUE self)