--- a/ext/liquid_c/tokenizer.h
+++ b/ext/liquid_c/tokenizer.h
@@ -13,6 +13,7 @@ typedef struct token {
     enum token_type type;
     const char *str;
     long length;
+    unsigned int trim_whitespace;
 } token_t;
 
 typedef struct tokenizer {
@@ -20,6 +21,7 @@ typedef struct tokenizer {
     const char *cursor;
     long length;
     unsigned int line_number;
+    unsigned int trim_whitespace;
 } tokenizer_t;
 
 extern VALUE cLiquidTokenizer;