--- a/ext/liquid_c/tokenizer.h
+++ b/ext/liquid_c/tokenizer.h
@@ -13,7 +13,8 @@ typedef struct token {
     enum token_type type;
     const char *str;
     long length;
-    unsigned int trim_whitespace;
+    unsigned int lstrip;
+    unsigned int rstrip;
 } token_t;
 
 typedef struct tokenizer {
@@ -21,7 +22,7 @@ typedef struct tokenizer {
     const char *cursor;
     long length;
     unsigned int line_number;
-    unsigned int trim_whitespace;
+    unsigned int lstrip_flag;
 } tokenizer_t;
 
 extern VALUE cLiquidTokenizer;