@evulse 2016-06-27
This adds support for {{- and {%- syntax which will lstrip! and -}} and -%} which will rstrip!