@evulse 2016-06-27
Now matches up to Shopify/liquid#773