@tobi 2016-06-27
I'd love to avoid the cycle through the Ruby VM to call rstrip!/lstrip!. 

Ideally we should parse white-space as a special token, then modify the token to be marked for render-skipping if {%- or -%} are encountered.