@tobi 2016-06-27
I see now that this would make it harder to match with Shopify/liquid#746 - alright, let's not overthink it. Thanks!