@tobi 2016-07-07
really nice work Mike.


- tobi
CEO Shopify

On Thu, Jul 7, 2016 at 8:18 AM, Mike Angell <notifications@github.com>
wrote:

> Alright we are passing again. Detection of the whitespace control
> characters has now been pushed back in to the tokenizer so there is no need
> to push and pop any existing values.
>
> —
> You are receiving this because you are on a team that was mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/Shopify/liquid-c/pull/30#issuecomment-231061305>, or mute
> the thread
> <https://github.com/notifications/unsubscribe/AAABW7m4s55YijtSESQGVQYhVRGXmrgdks5qTO6AgaJpZM4I_Yjh>
> .
>