@pushrax 2016-07-07
Definitely the better solution, nice.