@evulse 2016-07-08
Ok new change from using ruby rstrip is done which makes this pretty final. As such run the final performance benchmark

**Current**

    Calculating -------------------------------------
              parse:     80.921  (± 3.7%) i/s -      4.851k in  60.028478s
        parse & run:     32.196  (± 3.1%) i/s -      1.932k in  60.069380s

**This feature**

    Calculating -------------------------------------
              parse:     80.791  (± 3.7%) i/s -      4.844k in  60.031186s
        parse & run:     32.109  (± 3.1%) i/s -      1.926k in  60.046058s

The results show the performance impact of this feature is very minimal.