@dylanahsmith 2016-07-12
There were some corner cases that you missed as noted above.  But overall this is looking very good now.  Nice work