@evulse 2016-07-14
I think just shifting this removes this issue. So flowing through this previously this would have allowed `{.-` where . being any character. However i think shifting this only allows `{%-` and `{{-` as the tag. So the only flows are

- These 3 characters are at the end of a string and become a raw token
- These 3 characters are at the start and become a token or variable and run their own loops
- This is the last token and as such there will be no other tag following that will trigger the lstrip