@dylanahsmith 2016-07-18
It looks like https://github.com/Shopify/liquid-c/pull/30/files#r70563502 still hasn't been addressed.  That was my last concern.