@evulse 2016-07-20
Ok i've added a reset after this line so it will only be set if it is found.